Working on multiboot. Improve boot.s

+++ src/kernel/kernel.c
@@ -9,8 +9,8 @@
 
 // 内核入口
 // magic 魔数 addr multiboot 结构体的地址
-void kernel_main(uint32_t magic, uint32_t addr);
-void kernel_main(uint32_t magic, uint32_t addr) {
+void kernel_main(uint64_t magic, uint64_t addr);
+void kernel_main(uint64_t magic, uint64_t addr) {
 		// multiboot_info_t * mboot_ptr = (multiboot_info_t *)addr;
 		// 系统初始化
 		// debug_init(magic, mboot_ptr);
